onewire.py, ds18x20.py: Add support for parasitical power
This is through the method powermode() of the DS18X20 class, which
requires to supply a PIN object in OUT mode. Parasitic power will
be strengthened for conversion, at which time the Pin is set to 1.
A Pycom device GPIO port is sufficient. Connect that with a Schottky
diode like 1N5711 to the data pin is a minimal working set-up.